+++ config.py
@@ -1,7 +1,7 @@
 #coding:utf-8
 
 config = {
-	"VERSION":"2.4.3-beta5",
+	"VERSION":"2.4.3",
 	"WEB_API_VERSION":"0.4.3",
 	"localAddress":"127.0.0.1",
 	"localPort":1087,
